# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Waryaa Gaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List all tournaments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentInput"
      responses:
        "201":
          description: Created tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Not found
  /tournaments/{id}/register:
    post:
      operationId: registerForTournament
      tags: [tournaments]
      summary: Register a player or team for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentRegistration"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentParticipant"
  /tournaments/{id}/bracket:
    get:
      operationId: getTournamentBracket
      tags: [tournaments]
      summary: Get tournament bracket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bracket"
  /tournaments/{id}/matches:
    get:
      operationId: getTournamentMatches
      tags: [tournaments]
      summary: Get all matches for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
...
components:
  schemas:
...
    Tournament:
      type: object
      required: [id, name, status, format, maxParticipants, currentParticipants, prizePool]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        format:
          type: string
          enum: [single-elimination, double-elimination, round-robin, group-stage]
        game:
          type: string
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
        prizePool:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        streamUrl:
          type: ["string", "null"]
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        createdAt:
          type: string
    TournamentInput:
      type: object
      required: [name, format, maxParticipants, prizePool, startDate]
      properties:
        name:
          type: string
        description:
          type: string
        format:
          type: string
          enum: [single-elimination, double-elimination, round-robin, group-stage]
        game:
          type: string
        maxParticipants:
          type: integer
        prizePool:
          type: string
        startDate:
          type: string
        rules:
          type: string
        streamUrl:
          type: string
    TournamentRegistration:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [player, team]
        playerId:
          type: ["integer", "null"]
        teamId:
          type: ["integer", "null"]
    TournamentParticipant:
      type: object
      required: [id, tournamentId, type]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        type:
          type: string
        playerId:
          type: ["integer", "null"]
        playerName:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        seed:
          type: ["integer", "null"]
    Bracket:
      type: object
      required: [tournamentId, rounds]
      properties:
        tournamentId:
          type: integer
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/BracketRound"
    BracketRound:
      type: object
      required: [roundNumber, name, matches]
      properties:
        roundNumber:
          type: integer
        name:
          type: string
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"
    Match:
      type: object
      required: [id, tournamentId, round, status]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        round:
          type: integer
        roundName:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, live, completed, cancelled]
        participant1Id:
          type: ["integer", "null"]
        participant1Name:
          type: ["string", "null"]
        participant1Score:
          type: ["integer", "null"]
        participant2Id:
          type: ["integer", "null"]
        participant2Name:
          type: ["string", "null"]
        participant2Score:
          type: ["integer", "null"]
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        streamUrl:
          type: ["string", "null"]
        tournamentName:
          type: ["string", "null"]
...